# backend/resonand/api/routes/libraries.py
"""Libraries, their category tree and the grants on them (``API-8``).

Libraries are addressed by ``uuid``; categories are not, because they are only ever reached
through a library the caller can already read (``DEC-14``).

The share endpoints already accept the shape that individual-recording sharing will need. Only
library-level grants are offered in v0, and the endpoint that would grant one on a recording is
simply not routed -- the resolution behind it has been in place and tested since ``DAT-3``.
"""

from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Depends, status

from resonand.acl.query import require_library
from resonand.api.deps import (
    ArchiveDatabase,
    CurrentCaller,
    InstanceSettings,
    ReadSession,
    WriteSession,
)
from resonand.api.filters import RecordingFilters, RecordingSort
from resonand.api.pagination import Page, PageRequest, count_of, page_of, page_request
from resonand.api.presenters import (
    audio_summaries,
    category_summary,
    library_summaries,
    library_summary,
    share_summary,
)
from resonand.api.schemas import (
    AudioSummary,
    CategorySummary,
    CreateCategory,
    CreateLibrary,
    CreateShare,
    LibrarySummary,
    ReorderCategories,
    ShareSummary,
    UpdateCategory,
    UpdateLibrary,
)
from resonand.core.errors import NotFoundError
from resonand.core.levels import Level
from resonand.db import categories as category_repo
from resonand.db import libraries as library_repo
from resonand.db.audio import library_audio
from resonand.db.libraries import purge_library
from resonand.db.models import User
from resonand.media import storage

router = APIRouter(prefix="/libraries", tags=["libraries"])

Paging = Annotated[PageRequest, Depends(page_request)]


@router.get("", response_model=list[LibrarySummary], summary="Libraries you can see")
def list_libraries(caller: CurrentCaller, session: ReadSession) -> list[LibrarySummary]:
    """Owned libraries first, then shared ones -- which is the sidebar's own order (``UI-4``)."""
    return library_summaries(session, library_repo.list_libraries(session, caller.id))


@router.post("", response_model=LibrarySummary, status_code=status.HTTP_201_CREATED)
def create_library(
    body: CreateLibrary, caller: CurrentCaller, session: WriteSession
) -> LibrarySummary:
    library = library_repo.create_library(
        session, caller.id, name=body.name, description=body.description, colour=body.colour.value
    )
    return library_summary(session, library, Level.OWNER)


@router.get("/{library_uuid}", response_model=LibrarySummary)
def get_library(library_uuid: str, caller: CurrentCaller, session: ReadSession) -> LibrarySummary:
    library, level = require_library(session, caller.id, library_uuid)
    return library_summary(session, library, level)


@router.patch("/{library_uuid}", response_model=LibrarySummary)
def update_library(
    library_uuid: str, body: UpdateLibrary, caller: CurrentCaller, session: WriteSession
) -> LibrarySummary:
    library, level = library_repo.update_library(
        session,
        caller.id,
        library_uuid,
        name=body.name,
        description=body.description,
        colour=body.colour.value if body.colour is not None else None,
    )
    return library_summary(session, library, level)


@router.delete("/{library_uuid}", status_code=status.HTTP_204_NO_CONTENT)
def trash_library(library_uuid: str, caller: CurrentCaller, session: WriteSession) -> None:
    """Send a library to the trash. Nothing is deleted until the retention period runs out."""
    library_repo.trash_library(session, caller.id, library_uuid)


@router.post("/{library_uuid}/restore", response_model=LibrarySummary)
def restore_library(
    library_uuid: str, caller: CurrentCaller, session: WriteSession
) -> LibrarySummary:
    library, level = library_repo.restore_library(session, caller.id, library_uuid)
    return library_summary(session, library, level)


@router.get("/{library_uuid}/audio", response_model=Page[AudioSummary])
def list_library_audio(
    library_uuid: str,
    caller: CurrentCaller,
    session: ReadSession,
    paging: Paging,
    filters: RecordingFilters,
    sorting: RecordingSort,
) -> Page[AudioSummary]:
    """What is in a library: filtered, sorted, newest recording first by default (``API-10``).

    The filters are the same dependency search takes, and the sort fields are the ones the filter
    bar and the column headings both offer -- ``UI-7d`` calls those one sort expressed two ways,
    which is only true if there is one list of them.
    """
    sort, direction = sorting
    query = library_audio(
        session, caller.id, library_uuid, filters=filters, sort=sort, direction=direction
    )
    total = count_of(session, query)
    rows = session.execute(query.limit(paging.limit).offset(paging.offset)).all()
    return page_of(
        audio_summaries(session, [(row[0], row[1]) for row in rows]),
        total=total,
        request=paging,
    )


# --- Sharing --------------------------------------------------------------


@router.get("/{library_uuid}/shares", response_model=list[ShareSummary])
def list_shares(
    library_uuid: str, caller: CurrentCaller, session: ReadSession
) -> list[ShareSummary]:
    return [
        share_summary(share, grantee)
        for share, grantee in library_repo.list_shares(session, caller.id, library_uuid)
    ]


@router.put("/{library_uuid}/shares", response_model=ShareSummary)
def share_library(
    library_uuid: str, body: CreateShare, caller: CurrentCaller, session: WriteSession
) -> ShareSummary:
    """Grant access, or change the level somebody already has. Requires manage."""
    share = library_repo.share_library(
        session, caller.id, library_uuid, grantee_id=body.grantee_id, level=body.level
    )
    grantee = session.get(User, body.grantee_id)
    if grantee is None:  # pragma: no cover -- share_library already refused an unknown one
        raise NotFoundError("No such account.")
    return share_summary(share, grantee)


@router.delete("/{library_uuid}/shares/{grantee_id}", status_code=status.HTTP_204_NO_CONTENT)
def unshare_library(
    library_uuid: str, grantee_id: int, caller: CurrentCaller, session: WriteSession
) -> None:
    library_repo.unshare_library(session, caller.id, library_uuid, grantee_id=grantee_id)


# --- The category tree ----------------------------------------------------


@router.get("/{library_uuid}/categories", response_model=list[CategorySummary])
def list_categories(
    library_uuid: str, caller: CurrentCaller, session: ReadSession
) -> list[CategorySummary]:
    return [
        category_summary(category)
        for category in category_repo.list_categories(session, caller.id, library_uuid)
    ]


@router.post(
    "/{library_uuid}/categories",
    response_model=CategorySummary,
    status_code=status.HTTP_201_CREATED,
)
def create_category(
    library_uuid: str, body: CreateCategory, caller: CurrentCaller, session: WriteSession
) -> CategorySummary:
    return category_summary(
        category_repo.create_category(
            session, caller.id, library_uuid, name=body.name, parent_id=body.parent_id
        )
    )


@router.patch("/{library_uuid}/categories/{category_id}", response_model=CategorySummary)
def update_category(
    library_uuid: str,
    category_id: int,
    body: UpdateCategory,
    caller: CurrentCaller,
    session: WriteSession,
) -> CategorySummary:
    """Rename or re-parent one node. Moving it into its own subtree is refused."""
    require_library(session, caller.id, library_uuid, Level.EDIT)
    category = category_repo.get_category(session, caller.id, category_id, Level.EDIT)
    if body.name is not None:
        category = category_repo.rename_category(session, caller.id, category_id, name=body.name)
    if body.clear_parent or body.parent_id is not None:
        category = category_repo.move_category(
            session, caller.id, category_id, parent_id=None if body.clear_parent else body.parent_id
        )
    return category_summary(category)


@router.post("/{library_uuid}/categories/order", response_model=list[CategorySummary])
def reorder_categories(
    library_uuid: str,
    body: ReorderCategories,
    caller: CurrentCaller,
    session: WriteSession,
) -> list[CategorySummary]:
    require_library(session, caller.id, library_uuid, Level.EDIT)
    return [
        category_summary(category)
        for category in category_repo.reorder_categories(session, caller.id, body.ordered_ids)
    ]


@router.delete("/{library_uuid}/categories/{category_id}", status_code=status.HTTP_204_NO_CONTENT)
def delete_category(
    library_uuid: str, category_id: int, caller: CurrentCaller, session: WriteSession
) -> None:
    """Remove a node, uncategorising anything that was in it rather than refusing."""
    require_library(session, caller.id, library_uuid, Level.EDIT)
    category_repo.delete_category(session, caller.id, category_id)


trash_router = APIRouter(tags=["libraries"])
"""The trash lives outside the ``/libraries`` prefix, beside ``/trash/audio``.

``INT-1`` draws one list with a type marker rather than two sections -- the question somebody has
is where a thing went, not whether it was a library -- and the interface merges the two endpoints
client-side. Sitting them next to each other in the URL space is what makes that read as one
thing.
"""


@trash_router.delete("/trash/libraries/{library_uuid}", status_code=status.HTTP_204_NO_CONTENT)
def purge_trashed_library(
    library_uuid: str,
    caller: CurrentCaller,
    database: ArchiveDatabase,
    settings: InstanceSettings,
) -> None:
    """Destroy one trashed library now, with everything in it (``API-19``).

    It takes the recordings whether or not they were trashed separately, which is what trashing a
    library already means: the retention purge expires everything inside one on the library's own
    clock, so Delete now has to destroy the same set or it would leave behind exactly the rows
    waiting a month would have taken. ``INT-1c``'s confirmation states that count first.

    Its transaction is its own, for the reason the recording's purge gives.
    """
    with database.write_session() as session:
        removed = purge_library(session, caller.id, library_uuid)
    for audio_uuid in removed:
        storage.delete_recording(settings.resolved_storage_dir, audio_uuid)


@trash_router.get("/trash/libraries", response_model=Page[LibrarySummary])
def list_trashed_libraries(
    caller: CurrentCaller, session: ReadSession, paging: Paging
) -> Page[LibrarySummary]:
    """Trashed libraries, closest to being purged first, mirroring ``/trash/audio``.

    It takes level 30, the same as trashing one: somebody who could only read a library has no
    business being told it is on its way out.
    """
    query = library_repo.trashed_libraries(caller.id)
    total = count_of(session, query)
    rows = session.execute(query.limit(paging.limit).offset(paging.offset)).all()
    return page_of(
        library_summaries(session, [(row[0], row[1]) for row in rows]),
        total=total,
        request=paging,
    )