# frontend/src/api/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateLibrary": {
        "additionalProperties": false,
        "properties": {
          "colour": {
            "$ref": "#/components/schemas/Colour",
            "default": "stone"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateLibrary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SortDirection": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection",
        "type": "string"
      },
...
      "TranscriptionState": {
        "description": "What the badge says, and what the filter selects.",
        "enum": [
          "none",
          "running",
          "done",
          "failed"
        ],
        "title": "TranscriptionState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "A self-hosted archive for the recordings that matter.\n\nPublic identifiers are `uuid`s; integer ids are internal. Anything you may not read answers\n**404**, never 403 -- a 403 would confirm the resource exists.\n",
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "summary": "A self-hosted archive for the recordings that matter.",
    "title": "Resonand",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/libraries": {
      "get": {
        "description": "Owned libraries first, then shared ones -- which is the sidebar's own order (``UI-4``).",
        "operationId": "list_libraries_api_libraries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LibrarySummary"
                  },
                  "title": "Response List Libraries Api Libraries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Libraries you can see",
        "tags": [
          "libraries"
        ]
      },
      "post": {
        "operationId": "create_library_api_libraries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibrary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Library",
        "tags": [
          "libraries"
        ]
      }
    },
    "/api/libraries/{library_uuid}": {
      "delete": {
        "description": "Send a library to the trash. Nothing is deleted until the retention period runs out.",
        "operationId": "trash_library_api_libraries__library_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trash Library",
        "tags": [
          "libraries"
        ]
      },
      "get": {
        "operationId": "get_library_api_libraries__library_uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Library",
        "tags": [
          "libraries"
        ]
      },
      "patch": {
        "operationId": "update_library_api_libraries__library_uuid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibrary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Library",
        "tags": [
          "libraries"
        ]
      }
    },
    "/api/libraries/{library_uuid}/audio": {
      "get": {
        "description": "What is in a library: filtered, sorted, newest recording first by default (``API-10``).\n\nThe filters are the same dependency search takes, and the sort fields are the ones the filter\nbar and the column headings both offer -- ``UI-7d`` calls those one sort expressed two ways,\nwhich is only true if there is one list of them.",
        "operationId": "list_library_audio_api_libraries__library_uuid__audio_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "description": "Tag slugs, all of which must match.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tag slugs, all of which must match.",
              "title": "Tag"
            }
          },
          {
            "description": "Wall-clock, inclusive.",
            "in": "query",
            "name": "recorded_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Wall-clock, inclusive.",
              "title": "Recorded From"
            }
          },
          {
            "description": "Wall-clock, inclusive.",
            "in": "query",
            "name": "recorded_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Wall-clock, inclusive.",
              "title": "Recorded To"
            }
          },
          {
            "in": "query",
            "name": "min_duration_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Duration Ms"
            }
          },
          {
            "in": "query",
            "name": "max_duration_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Duration Ms"
            }
          },
          {
            "description": "Any of the four states. Repeat it to mean either.",
            "in": "query",
            "name": "transcription_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/TranscriptionState"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any of the four states. Repeat it to mean either.",
              "title": "Transcription State"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortField",
              "default": "recorded_at"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDirection",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AudioSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Library Audio",
        "tags": [
          "libraries"
        ]
      },
      "post": {
        "description": "Store a recording and queue the work that turns it into an archived one.\n\nThe bytes are written and hashed in one pass -- an hours-long upload is not walked twice --\nand the original is never rewritten afterwards.\n\n**The database is not touched while they are arriving** (``REV-1``). The identifier is minted\nhere rather than by the insert, so the file can be written under its final name with nothing\nopen, and the row that describes it goes in afterwards in one short transaction. An upload\nthat held the write lock would stop every other writer in the instance -- the worker\nincluded -- for as long as the upload took, which for an hour of driving is an hour.\n\nThe permission is resolved twice deliberately: once before a byte is accepted, so a stranger\nis refused rather than served eight gigabytes of patience, and once inside the transaction\nthat writes, which is the one that decides.",
        "operationId": "upload_api_libraries__library_uuid__audio_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_libraries__library_uuid__audio_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a recording",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/libraries/{library_uuid}/categories": {
      "get": {
        "operationId": "list_categories_api_libraries__library_uuid__categories_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategorySummary"
                  },
                  "title": "Response List Categories Api Libraries  Library Uuid  Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Categories",
        "tags": [
          "libraries"
        ]
      },
      "post": {
        "operationId": "create_category_api_libraries__library_uuid__categories_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Category",
        "tags": [
          "libraries"
        ]
      }
    },
    "/api/libraries/{library_uuid}/categories/order": {
      "post": {
        "operationId": "reorder_categories_api_libraries__library_uuid__categories_order_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCategories"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategorySummary"
                  },
                  "title": "Response Reorder Categories Api Libraries  Library Uuid  Categories Order Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder Categories",
        "tags": [
          "libraries"
        ]
      }
    },
    "/api/libraries/{library_uuid}/categories/{category_id}": {
      "delete": {
        "description": "Remove a node, uncategorising anything that was in it rather than refusing.",
        "operationId": "delete_category_api_libraries__library_uuid__categories__category_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Category",
        "tags": [
          "libraries"
        ]
      },
      "patch": {
        "description": "Rename or re-parent one node. Moving it into its own subtree is refused.",
        "operationId": "update_category_api_libraries__library_uuid__categories__category_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Category",
        "tags": [
          "libraries"
        ]
      }
    },
    "/api/libraries/{library_uuid}/restore": {
      "post": {
        "operationId": "restore_library_api_libraries__library_uuid__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Library",
        "tags": [
          "libraries"
        ]
      }
    },
...
    "/api/libraries/{library_uuid}/shares/{grantee_id}": {
      "delete": {
        "operationId": "unshare_library_api_libraries__library_uuid__shares__grantee_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "title": "Library Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "grantee_id",
            "required": true,
            "schema": {
              "title": "Grantee Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unshare Library",
        "tags": [
          "libraries"
        ]
      }
    },
...
  }
}